+++ site/AntDesign.Docs.Build.CLI/Documentations/PropertyDocumentation.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace AntDesign.Docs.Build.CLI
+namespace AntDesign.Docs.Build.CLI.Documentations
 {
     public class PropertyDocumentation : IApiDocumentation
     {
