+++ site/AntDesign.Docs.Build.CLI/Services/Translation/AzureTranslateOptions.cs
